pyproject.toml
[tool.poetry]
name = "sheets_api"
version = "0.1.0"
description = "test homework"
authors = ["Liudmla Patsukevich <devbrest@gmail.com>"]

[tool.poetry.scripts]
exchange = "sheets_api.exchange:run_script"
running_script = "sheets_api.exchange:start_app"
create_database = "sheets_api.data:create_database"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.3"
pandas = "^1.3.4"
numpy = "^0.0"
psycopg2 = "^2.7"
google-api-python-client = "^0.0"
google-auth-httplib2 = "^0.0"
google-auth-oauthlib = "^0.0"
oauth2client = "^4.1.3"

[tool.poetry.extras]
pgsql = ["psycopg2"]

[tool.poetry.dev-dependencies]
black = "^21.11b1"
mypy = "^0.910"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
